+++ controllers/rebuildfile.go
@@ -7,23 +7,11 @@ import (
 	"net/http"
 
 	"github.com/k8-proxy/k8-go-api/models"
+	"github.com/k8-proxy/k8-go-api/pkg/store"
 	"github.com/k8-proxy/k8-go-api/utils"
-	"github.com/streadway/amqp"
 
+	"github.com/k8-proxy/k8-go-api/pkg/message"
 	"github.com/rs/zerolog"
-
-	"github.com/k8-proxy/k8-go-api/pkg/minio"
-	"github.com/k8-proxy/k8-go-api/pkg/rabbitmq"
-)
-
-var (
-	exchange   = "adaptation-exchange"
-	routingKey = "adaptation-request"
-	queueName  = "adaptation-request-queue"
-
-	processing_exchange   = "processing-exchange"
-	processing_routingKey = "processing-request"
-	processing_queueName  = "processing-queue"
 )
 
 // RebuildFile rebuilds a file using its binary data
@@ -69,63 +57,8 @@ func RebuildFile(w http.ResponseWriter, r *http.Request) {
 	})
 
 	/////////////////////////////
-	minioEndpoint := "localhost:9000"
-	minioAccessKey := "minioadmin"
-	minioSecretKey := "minioadmin"
-	sourceMinioBucket := "test"
-
-	cl := minio.NewMinioClient(minioEndpoint, minioAccessKey, minioSecretKey, false)
-	exist, err := minio.CheckIfBucketExists(cl, sourceMinioBucket)
-	if err != nil || !exist {
-		log.Println("error checkbucket ", err)
-		utils.ResponseWithError(w, http.StatusInternalServerError, "bucket not exist")
-		return
-	}
-
-	publisher, err := rabbitmq.NewQueuePublisher(conn, exchange)
-	if err != nil {
-		log.Fatalf("%s", err)
-	}
-
-	defer publisher.Close()
-
-	// Start a consumer
-	msgs, ch, err := rabbitmq.NewQueueConsumer(conn, processing_queueName, processing_exchange, processing_routingKey)
-	if err != nil {
-		log.Fatalf("%s", err)
-	}
-	defer ch.Close()
-
-	table := amqp.Table{
-		"file-id":               "myfileid",
-		"source-file-location":  "/home/ibrahim/my_work/k8-go-api/sampledata/file.pdf",
-		"rebuilt-file-location": "sampledata/file.zip",
-	}
-
-	err = rabbitmq.PublishMessage(publisher, exchange, routingKey, table, []byte("ibrahim"))
-	if err != nil {
-		log.Println("PublishMessage", err)
-
-		return
-	}
-
-	var miniourl string
-
-	notforever := make(chan bool)
-
-	go func() {
-		for d := range msgs {
-			log.Printf("Received a message: %s", d.Body)
-			miniourl = d.Headers["source-presigned-url"].(string)
-			notforever <- true
-		}
-	}()
-	<-notforever
-
-	log.Printf(" [*] Waiting for messages. To exit press CTRL+C")
-
-	log.Println(miniourl)
-
+	store.St()
+	message.AmqpM()
 	/////////////////////////
 	//GW custom header
 	utils.AddGWHeader(w, models.Temp)
